+++ src/app/layout.tsx
@@ -27,13 +27,6 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <head>
-        {/* eslint-disable-next-line @next/next/no-page-custom-font */}
-        <link
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:wght,FILL@100..700,0..1&display=swap"
-          rel="stylesheet"
-        />
-      </head>
       <body className={`${manrope.className} ${manrope.variable} ${dmSerif.variable} antialiased`}>
         <AppShell navSections={navSections}>{children}</AppShell>
       </body>